# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Book-A-Meal is an application that allows customers to make food orders and helps the food  vendor know what the customers want to eat.",
    "version": "1.0.0",
    "title": "Book A Meal",
    "contact": {
      "email": "phemy.smith@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://book-meal.herokuapp.com/api/v1"
    }
  },
  "host": "book-meal.herokuapp.com",
  "basePath": "/api/v1",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/meals": {
      "post": {
        "tags": [
          "Meals"
        ],
        "summary": "Creates new meal",
        "description": "This method is called to add a new meal",
        "operationId": "createmeal",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Authentication Token",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "creating a new meal",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Meals"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Meal successfull created"
          },
          "400": {
            "description": "input errors",
            "schema": {
              "$ref": "#/definitions/ResponseMessage"
            }
          },
          "409": {
            "description": "meal already exist",
            "schema": {
              "$ref": "#/definitions/ResponseMessage"
            }
          },
          "500": {
            "description": "meal cannot be created",
            "schema": {
              "$ref": "#/definitions/ResponseMessage"
            }
          },
          "default": {
            "description": "Unexpected Error"
          }
        }
      },
      "get": {
        "tags": [
          "Meals"
        ],
        "summary": "Get all meals",
        "description": "gets all meals created",
        "operationId": "getmeal",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Authentication Token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Meals"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/ResponseMessage"
            }
          },
          "default": {
            "description": "Unexpected Error"
          }
        }
      }
    },
    "/meals/<mealId>": {
      "put": {
        "tags": [
          "Meals"
        ],
        "summary": "Modify an existing meal",
        "description": "",
        "operationId": "editmeal",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Authentication Token",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "mealId",
            "required": true,
            "description": "ID of meal to be modified",
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "food form model",
            "schema": {
              "$ref": "#/definitions/Meals"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meal successfully updated"
          },
          "400": {
            "description": "fail to modify meal, Invalid input"
          },
          "404": {
            "description": "Meal not found"
          },
          "500": {
            "description": "Meal cannot be updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Meals"
        ],
        "summary": "delete an existing meal",
        "description": "",
        "operationId": "deletemeal",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Authentication Token",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "mealId",
            "required": true,
            "description": "ID of meal to be deleted",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "meal successfully deleted"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Meal not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "book_a_meal_auth": [
              "write:meals",
              "read:meals"
            ]
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "book_a_meal_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://book-meal.herokuapp.com/api/v1/auth",
      "flow": "implicit",
      "scopes": {
        "write:meals": "modify a meal",
        "read:meals": "read a meal",
        "write:orders": "create/modify orders",
        "read:orders": "read orders",
        "write:menu": "create/modify a menu",
        "read:menu": "read a menu"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "Meals": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "rice"
        },
        "description": {
          "type": "string",
          "example": "a very delicious meal"
        },
        "price": {
          "type": "number",
          "example": 500
        },
        "image": {
          "type": "string",
          "example": "https://cdn-images-1.medium.com/max/1600/1*CZ3OBsPfLV8XqKKn7Aii5Q.png"
        }
      }
    },
...
    "ResponseMessage": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  },
...
}